Add files via upload
With the technical advancement in the medical
discipline, early detection of most of the diseases has
become practicable. Our research work proposes a CNN
application which can detect and classify skin cancer from
raw Dermoscopic Images. Early detection and classification
of skin cancer lesion can significantly reduce the diagnosis
period which increases the odds of recovery. Classification
models used to detect skin cancer lesions and techniques
involved pertaining to them are discussed in the related work
section. In addition to the work done in the current state
of art model, we augmented the images of the classes which
had the minor data segment and were responsible for the
class imbalance. In order to feed the noise-free images to the
model, we pre-processed the images by performing iterations
of erosion and dilution operations. Our research was modelled
on two convolutions layered Neural Network and achieved the
best validation accuracy of 89.49 per cent. Our experiment
output was marginally less than the current state of the art model
(89.99) in skin cancer lesion classification.